+++ now-waypoint-core/src/main/java/isn_t_this_e_not_i/now_waypoint_core/domain/auth/exception/auth/NullFieldException.java
@@ -1,4 +1,4 @@
-package isn_t_this_e_not_i.now_waypoint_core.domain.auth.exception;
+package isn_t_this_e_not_i.now_waypoint_core.domain.auth.exception.auth;
 
 public class NullFieldException extends RuntimeException{
     public NullFieldException(String message) {
